Extract component to render a list of tags
