+++ app/components/shared/tag_list_component.html.erb
@@ -1,0 +1,1 @@
+<%= link_list(*links, class: "tags", id: "tags_#{dom_id(taggable)}") %>
